Support locals in provider blocks
