+++ pkg/tf2pulumi/convert/testdata/provider_config_locals/experimental_pcl/Pulumi.yaml
@@ -1,0 +1,5 @@
+name: provider_config_locals
+runtime: terraform
+config:
+    configured:stringConfig:
+        value: Static
