+++ pkg/tf2pulumi/convert/testdata/provider_config_locals/experimental_pcl/main.pp
@@ -1,0 +1,7 @@
+// Check we can use terraform builtin functions here and that they are evaluated
+staticLocal = invoke("std:index:title", {
+  input = "static"
+}).result
+resource "aDefaultResource" "configured:index:resource" {
+  inputOne = staticLocal
+}
