+++ pkg/tf2pulumi/convert/testdata/provider_config_vars/experimental_pcl/Pulumi.yaml
@@ -1,0 +1,5 @@
+name: provider_config_vars
+runtime: terraform
+config:
+    configured:stringConfig:
+        value: 'TODO: var.region'
