+++ pkg/tf2pulumi/convert/testdata/provider_config_vars/experimental_pcl/main.pp
@@ -1,0 +1,6 @@
+config "region" "string" {
+  description = "The region to use"
+}
+resource "aDefaultResource" "configured:index:resource" {
+  inputOne = region
+}
