+++ pkg/tf2pulumi/convert/tf.go
@@ -32,8 +32,6 @@ import (
 	"github.com/hashicorp/terraform-svchost/disco"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tf2pulumi/il"
 	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfbridge"
-	shim "github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim"
-	"github.com/pulumi/pulumi-terraform-bridge/v3/pkg/tfshim/schema"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/hcl2/syntax"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/pcl"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/tokens"
@@ -688,12 +686,17 @@ var tfFunctionStd = map[string]struct {
 	},
 }
 
-// Returns a call to notImplemented with the text of the input range, e.g. `notImplemented("some.expr[0]")`
-func notImplemented(sources map[string][]byte, rng hcl.Range) hclwrite.Tokens {
+// Returns the source code for the given range
+func sourceCode(sources map[string][]byte, rng hcl.Range) string {
 	buffer := bytes.NewBufferString("")
 	_, err := getTokensForRange(sources, rng).WriteTo(buffer)
 	contract.AssertNoErrorf(err, "Failed to write tokens for range %v", rng)
-	text := cty.StringVal(strings.Replace(buffer.String(), "\r\n", "\n", -1))
+	return strings.Replace(buffer.String(), "\r\n", "\n", -1)
+}
+
+// Returns a call to notImplemented with the text of the input range, e.g. `notImplemented("some.expr[0]")`
+func notImplemented(sources map[string][]byte, rng hcl.Range) hclwrite.Tokens {
+	text := cty.StringVal(sourceCode(sources, rng))
 	return hclwrite.TokensForFunctionCall("notImplemented", hclwrite.TokensForValue(text))
 }
 
@@ -2102,350 +2105,6 @@ func convertOutput(sources map[string][]byte, scopes *scopes,
 	return leading, block, trailing
 }
 
-// Used to return info about a path in the schema.
-type PathInfo struct {
-	// The final part of the path (e.g. a_field)
-	Name string
-
-	// The Resource that contains the path (e.g. data.simple_data_source)
-	Resource shim.Resource
-	// The DataSourceInfo for the path (e.g. data.simple_data_source)
-	DataSourceInfo *tfbridge.DataSourceInfo
-	// The ResourceInfo for the path (e.g. simple_resource)
-	ResourceInfo *tfbridge.ResourceInfo
-
-	// The Schema for the path (e.g. data.simple_data_source.a_field)
-	Schema shim.Schema
-	// The SchemaInfo for the path (e.g. data.simple_data_source.a_field)
-	SchemaInfo *tfbridge.SchemaInfo
-}
-
-type scopes struct {
-	info il.ProviderInfoSource
-
-	// All known roots, keyed by fully qualified path e.g. data.some_data_source
-	roots map[string]PathInfo
-
-	// Local variables that are in scope from for expressions
-	locals []map[string]string
-
-	// Set non-nil if "count.index" can be mapped
-	countIndex hcl.Traversal
-	eachKey    hcl.Traversal
-	eachValue  hcl.Traversal
-}
-
-// lookup the given name in roots and locals
-func (s *scopes) lookup(name string) string {
-	for i := len(s.locals) - 1; i >= 0; i-- {
-		if s.locals[i][name] != "" {
-			return s.locals[i][name]
-		}
-	}
-	if root, has := s.roots[name]; has {
-		return root.Name
-	}
-	return ""
-}
-
-func (s *scopes) push(locals map[string]string) {
-	s.locals = append(s.locals, locals)
-}
-
-func (s *scopes) pop() {
-	s.locals = s.locals[0 : len(s.locals)-1]
-}
-
-// isUsed returns if _any_ root scope currently uses the name "name"
-func (s *scopes) isUsed(name string) bool {
-	// We don't have many, but there's a few _keywords_ in pcl that are easier if we just never emit them
-	if name == "range" {
-		return true
-	}
-
-	for _, usedName := range s.roots {
-		if usedName.Name == name {
-			return true
-		}
-	}
-	return false
-}
-
-// generateUniqueName takes "name" and ensures it's unique.
-// First by appending `suffix` to it, and then appending an incrementing count
-func (s *scopes) generateUniqueName(name, prefix, suffix string) string {
-	// Not used, just return it
-	if !s.isUsed(name) {
-		return name
-	}
-	// It's used, so add the prefix and suffix
-	name = prefix + name + suffix
-	if !s.isUsed(name) {
-		return name
-	}
-	// Still used add a counter
-	baseName := name
-	counter := 2
-	for {
-		name = fmt.Sprintf("%s%d", baseName, counter)
-		if !s.isUsed(name) {
-			return name
-		}
-		counter = counter + 1
-	}
-}
-
-func (s *scopes) getOrAddOutput(name string) string {
-	root, has := s.roots[name]
-	if has {
-		return root.Name
-	}
-	parts := strings.Split(name, ".")
-	tfName := parts[len(parts)-1]
-	pulumiName := camelCaseName(tfName)
-	s.roots[name] = PathInfo{Name: pulumiName}
-	return pulumiName
-}
-
-// getPulumiName takes "name" and ensures it's unique.
-// First by appending `suffix` to it, and then appending an incrementing count
-func (s *scopes) getOrAddPulumiName(name, prefix, suffix string) string {
-	root, has := s.roots[name]
-	if has {
-		return root.Name
-	}
-	parts := strings.Split(name, ".")
-	tfName := parts[len(parts)-1]
-	pulumiName := camelCaseName(tfName)
-	pulumiName = s.generateUniqueName(pulumiName, prefix, suffix)
-	s.roots[name] = PathInfo{Name: pulumiName}
-	return pulumiName
-}
-
-// Given a fully typed path (e.g. data.simple_data_source.a_field) returns the final part of that path
-// (a_field) and the either the Resource or Schema, and SchemaInfo for that path (if any).
-func (s *scopes) getInfo(fullyQualifiedPath string) PathInfo {
-	parts := strings.Split(fullyQualifiedPath, ".")
-	contract.Assertf(len(parts) >= 2, "empty path passed into getInfo: %s", fullyQualifiedPath)
-	contract.Assertf(parts[0] != "", "empty path part passed into getInfo: %s", fullyQualifiedPath)
-	contract.Assertf(parts[1] != "", "empty path part passed into getInfo: %s", fullyQualifiedPath)
-
-	var getInner func(sch shim.SchemaMap, info map[string]*tfbridge.SchemaInfo, parts []string) PathInfo
-	getInner = func(sch shim.SchemaMap, info map[string]*tfbridge.SchemaInfo, parts []string) PathInfo {
-		contract.Assertf(parts[0] != "", "empty path part passed into getInfo")
-
-		// At this point parts[0] may be an property + indexer or just a property. Work that out first.
-		part, rest, indexer := strings.Cut(parts[0], "[]")
-
-		// Lookup the info for this part
-		var curSch shim.Schema
-		if sch != nil {
-			curSch = sch.Get(part)
-		}
-		curInfo := info[part]
-
-		// We want this part
-		if len(parts) == 1 && !indexer {
-			return PathInfo{
-				Name:       part,
-				Schema:     curSch,
-				SchemaInfo: curInfo,
-			}
-		}
-
-		// Else recurse into the next part of the type, how we do this depends on if this was indexed or not
-		if !indexer {
-			// No indexers, simple recurse on fields
-			var nextSchema shim.SchemaMap
-			var nextInfo map[string]*tfbridge.SchemaInfo
-			if curSch != nil {
-				if sch, ok := curSch.Elem().(shim.Resource); ok {
-					nextSchema = sch.Schema()
-				}
-			}
-			if curInfo != nil {
-				nextInfo = curInfo.Fields
-			}
-			return getInner(nextSchema, nextInfo, parts[1:])
-		}
-
-		// part was indexed (i.e. something like "part[]" or part[]foo[][]bar), so rather than looking at the
-		// fields we need to look at the elements.
-		for {
-			var resourceOrSchema interface{}
-			if curSch != nil {
-				resourceOrSchema = curSch.Elem()
-			}
-			if curInfo != nil {
-				curInfo = curInfo.Elem
-			}
-
-			if rest == "" && len(parts) == 1 {
-				// This element is what we we're looking for
-				res, _ := resourceOrSchema.(shim.Resource)
-				sch, _ := resourceOrSchema.(shim.Schema)
-				return PathInfo{
-					Name:       part,
-					Resource:   res,
-					Schema:     sch,
-					SchemaInfo: curInfo,
-				}
-			} else if rest == "" {
-				// Can recurse into the next set of fields
-				var nextSchema shim.SchemaMap
-				var nextInfo map[string]*tfbridge.SchemaInfo
-				if sch, ok := resourceOrSchema.(shim.Resource); ok {
-					nextSchema = sch.Schema()
-				}
-				if curInfo != nil {
-					nextInfo = curInfo.Fields
-				}
-				return getInner(nextSchema, nextInfo, parts[1:])
-			}
-
-			panic(fmt.Sprintf("complex indexerParts not implemented: %v", rest))
-		}
-	}
-
-	if parts[0] == "data" {
-		contract.Assertf(len(parts) >= 3, "empty path passed into getInfo: %s", fullyQualifiedPath)
-		contract.Assertf(parts[2] != "", "empty path part passed into getInfo: %s", fullyQualifiedPath)
-
-		root, has := s.roots[parts[0]+"."+parts[1]+"."+parts[2]]
-		if len(parts) == 3 {
-			if has {
-				return root
-			}
-			// If we don't have a root, just return the name
-			return PathInfo{Name: parts[2]}
-		}
-
-		var currentSchema shim.SchemaMap
-		var currentInfo map[string]*tfbridge.SchemaInfo
-		if root.Resource != nil {
-			currentSchema = root.Resource.Schema()
-		}
-		if root.DataSourceInfo != nil {
-			currentInfo = root.DataSourceInfo.Fields
-		}
-
-		return getInner(currentSchema, currentInfo, parts[3:])
-
-	}
-
-	root, has := s.roots[parts[0]+"."+parts[1]]
-
-	if len(parts) == 2 {
-		if has {
-			return root
-		}
-		// If we don't have a root, just return the name
-		return PathInfo{Name: parts[1]}
-	}
-
-	var currentSchema shim.SchemaMap
-	var currentInfo map[string]*tfbridge.SchemaInfo
-	if root.Resource != nil {
-		currentSchema = root.Resource.Schema()
-	}
-	if root.ResourceInfo != nil {
-		currentInfo = root.ResourceInfo.Fields
-	}
-
-	return getInner(currentSchema, currentInfo, parts[2:])
-}
-
-// Given a fully typed path (e.g. data.simple_data_source.my_data.a_field) returns the pulumi name for that path.
-func (s *scopes) pulumiName(fullyQualifiedPath string) string {
-	info := s.getInfo(fullyQualifiedPath)
-
-	// This should only be called for attribute paths, so panic if this returned a resource
-	contract.Assertf(info.ResourceInfo == nil, "pulumiName called on a resource or data source")
-	contract.Assertf(info.DataSourceInfo == nil, "pulumiName called on a resource or data source")
-
-	// If we have a SchemaInfo and name use it
-	schemaInfo := info.SchemaInfo
-	if schemaInfo != nil && schemaInfo.Name != "" {
-		return schemaInfo.Name
-	}
-
-	// If we have a shim schema use it to translate
-	sch := info.Schema
-	if sch != nil {
-		return tfbridge.TerraformToPulumiNameV2(info.Name,
-			schema.SchemaMap(map[string]shim.Schema{info.Name: sch}),
-			map[string]*tfbridge.SchemaInfo{info.Name: schemaInfo})
-	}
-
-	// Else just return the name camel cased
-	return camelCaseName(info.Name)
-}
-
-// Given a fully typed path (e.g. data.simple_data_source.my_data.a_field) returns if the schema says it's a map.
-func (s *scopes) isMap(fullyQualifiedPath string) *bool {
-	info := s.getInfo(fullyQualifiedPath)
-
-	// This should only be called for attribute paths, so panic if this returned a resource
-	contract.Assertf(info.ResourceInfo == nil, "isMap called on a resource or data source")
-	contract.Assertf(info.DataSourceInfo == nil, "isMap called on a resource or data source")
-
-	// If we have a shim schema use the type from that
-	sch := info.Schema
-	if sch != nil {
-		isMap := sch.Type() == shim.TypeMap
-		return &isMap
-	}
-
-	// If we have a Resource schema this must be an object
-	if info.Resource != nil {
-		isMap := false
-		return &isMap
-	}
-
-	return nil
-}
-
-// Given a fully typed path (e.g. data.simple_data_source.a_field) returns whether a_field has maxItemsOne set
-func (s *scopes) maxItemsOne(fullyQualifiedPath string) bool {
-	info := s.getInfo(fullyQualifiedPath)
-
-	// This should only be called for attribute paths, so panic if this returned a resource
-	contract.Assertf(info.ResourceInfo == nil, "maxItemsOne called on a resource or data source")
-	contract.Assertf(info.DataSourceInfo == nil, "maxItemsOne called on a resource or data source")
-
-	// If we have a SchemaInfo and a MaxItems override use it
-	schemaInfo := info.SchemaInfo
-	if schemaInfo != nil && schemaInfo.MaxItemsOne != nil {
-		return *schemaInfo.MaxItemsOne
-	}
-
-	// If we have a shim schema use it's MaxItems
-	sch := info.Schema
-	if sch != nil {
-		return sch.MaxItems() == 1
-	}
-
-	// Else assume false
-	return false
-}
-
-// Given a fully typed path (e.g. data.simple_data_source.a_field) returns whether a_field has Asset information set
-func (s *scopes) isAsset(fullyQualifiedPath string) *tfbridge.AssetTranslation {
-	info := s.getInfo(fullyQualifiedPath)
-
-	// This should only be called for attribute paths, so panic if this returned a resource
-	contract.Assertf(info.ResourceInfo == nil, "isAsset called on a resource or data source")
-	contract.Assertf(info.DataSourceInfo == nil, "isAsset called on a resource or data source")
-
-	// If we have a SchemaInfo and a asset info return that
-	schemaInfo := info.SchemaInfo
-	if schemaInfo != nil {
-		return schemaInfo.Asset
-	}
-
-	return nil
-}
-
 // An "item" from a terraform file
 type terraformItem struct {
 	variable   *configs.Variable
@@ -2586,11 +2245,7 @@ func translateModuleSourceCode(
 		return moduleDiagnostics
 	}
 
-	scopes := &scopes{
-		info:   info,
-		roots:  make(map[string]PathInfo),
-		locals: make([]map[string]string, 0),
-	}
+	scopes := newScopes(info)
 
 	diagnostics := hcl.Diagnostics{}
 
@@ -2628,7 +2283,11 @@ func translateModuleSourceCode(
 	}
 	for _, item := range items {
 		if item.local != nil {
-			scopes.getOrAddPulumiName("local."+item.local.Name, "my", "")
+			key := "local." + item.local.Name
+			scopes.getOrAddPulumiName(key, "my", "")
+			root := scopes.roots[key]
+			root.Expression = &item.local.Expr
+			scopes.roots[key] = root
 		}
 	}
 	for _, item := range items {
@@ -2969,16 +2628,16 @@ func translateModuleSourceCode(
 			for _, attrKey := range attrKeys {
 				// Evauluate and marshal the attribute to a YAML like value for Pulumi config
 				value := attrs[attrKey]
-				val, diags := value.Expr.Value(nil)
-				diagnostics = append(diagnostics, diags...)
+				val, diags := scopes.EvalExpr(value.Expr)
 				if diags.HasErrors() {
 					diagnostics = append(diagnostics, &hcl.Diagnostic{
 						Subject:  &provider.DeclRange,
-						Severity: hcl.DiagError,
+						Severity: hcl.DiagWarning,
 						Summary:  "Failed to evaluate provider config",
-						Detail:   fmt.Sprintf("Failed to evaluate provider config for %s:%s", provider.Name, attrKey),
+						Detail:   fmt.Sprintf("Could not evaluate expression for %s:%s", provider.Name, attrKey),
 					})
-					continue
+					// If we couldn't eval the config we'll emit an obvious TODO to the config for it
+					val = cty.StringVal("TODO: " + sourceCode(sources, value.Expr.Range()))
 				}
 
 				// Simplest way to get a cty type into YAML is to roundtrip it through JSON
@@ -2988,7 +2647,7 @@ func translateModuleSourceCode(
 						Subject:  &provider.DeclRange,
 						Severity: hcl.DiagError,
 						Summary:  "Failed to marshal provider config",
-						Detail:   fmt.Sprintf("Failed to marshal provider config for %s:%s: %v", provider.Name, attrKey, err),
+						Detail:   fmt.Sprintf("Could not marshal value for %s:%s: %v", provider.Name, attrKey, err),
 					})
 					continue
 				}
